requirements.txt
# Core Streamlit dependencies
streamlit>=1.32.0
streamlit-extras>=0.3.5

# Essential ML dependencies for contract analysis
transformers>=4.30.0
torch>=2.5.0
sentence-transformers>=2.2.2
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# File processing
beautifulsoup4>=4.12.2
requests>=2.31.0
pillow>=10.1.0
pymupdf>=1.23.0
python-docx>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
jinja2>=3.1.0

